organize examples
