+++ examples/playground/lib/server/assertNotBrowser.ts
