+++ package.json
@@ -61,6 +61,8 @@
     "nodemon": "^2.0.2",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.0.5",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
     "ts-jest": "^25.2.1",
     "ts-node": "^9.1.1",
     "typescript": "^4.1.0-beta",
